pyproject.toml
[project]
name = "insurance-claim-negotiator"
version = "0.1.0"
description = "Voice-driven insurance claim settlement negotiator using Portia SDK"
requires-python = ">=3.11"

dependencies = [
    "portia-sdk-python>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.4.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.7",
    "httpx>=0.25.0",
    "asyncio-mqtt>=0.13.0",
    "websockets>=12.0",
    # Hume AI SDK
    "hume>=0.4.0",
    # Audio processing
    "pyaudio>=0.2.11",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    # Data processing
    "pandas>=2.1.0",
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.9.0",
    "mypy>=1.6.0",
    "pre-commit>=3.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
